lab12_homework/pyproject.toml
[project]
name = "MLOps_course_AGH_lab_02"
version = "1.0.0"
description = ""
readme = "README.md"

requires-python = "==3.11.*"

dependencies = [
    "apache-airflow[celery,postgres,standard,uv,s3fs]==3.1.5",
    "apache-airflow-providers-amazon[s3fs]",
    "twelvedata",
]

[dependency-groups]
dev = [
    "ruff"
]

airflow_common = [
    "apache-airflow[celery,postgres,standard,uv,s3fs]==3.1.5",
    "apache-airflow-providers-amazon[s3fs]",
    "requests",
    "s3fs",
]
